+++ src/CleanRegex/Exception/NoSuchStreamElementException.php
@@ -1,7 +1,7 @@
 <?php
 namespace TRegx\CleanRegex\Exception;
 
-class NoSuchElementFluentException extends \Exception implements PatternException
+class NoSuchStreamElementException extends \Exception implements PatternException
 {
     public function __construct(string $message)
     {