+++ test/Feature/CleanRegex/Match/asInt/AbstractMatchPatternTest.php
@@ -7,8 +7,8 @@ use Test\Fakes\CleanRegex\Internal\ThrowSubject;
 use Test\Utils\CustomSubjectException;
 use Test\Utils\ExactExceptionMessage;
 use Test\Utils\Functions;
-use TRegx\CleanRegex\Exception\NoSuchElementFluentException;
 use TRegx\CleanRegex\Exception\NoSuchNthElementException;
+use TRegx\CleanRegex\Exception\NoSuchStreamElementException;
 use TRegx\CleanRegex\Exception\SubjectNotMatchedException;
 use TRegx\CleanRegex\Internal\Match\Stream\StramRejectedException;
 
@@ -63,7 +63,7 @@ class AbstractMatchPatternTest extends TestCase
     public function shouldThrow_map_first_OnUnmatchedSubject()
     {
         // then
-        $this->expectException(NoSuchElementFluentException::class);
+        $this->expectException(NoSuchStreamElementException::class);
         $this->expectExceptionMessage('Expected to get the first match as integer, but subject was not matched');
 
         // when
@@ -76,7 +76,7 @@ class AbstractMatchPatternTest extends TestCase
     public function shouldThrow_map_nth_OnUnmatchedSubject()
     {
         // then
-        $this->expectException(NoSuchElementFluentException::class);
+        $this->expectException(NoSuchStreamElementException::class);
         $this->expectExceptionMessage('Expected to get the 0-nth stream element, but the subject backing the stream was not matched');
 
         // when
@@ -89,7 +89,7 @@ class AbstractMatchPatternTest extends TestCase
     public function shouldThrow_filter_first()
     {
         // then
-        $this->expectException(NoSuchElementFluentException::class);
+        $this->expectException(NoSuchStreamElementException::class);
         $this->expectExceptionMessage("Expected to get the first stream element, but the stream has 0 element(s)");
 
         // when
@@ -102,7 +102,7 @@ class AbstractMatchPatternTest extends TestCase
     public function shouldThrow_filter_nth()
     {
         // then
-        $this->expectException(NoSuchElementFluentException::class);
+        $this->expectException(NoSuchStreamElementException::class);
         $this->expectExceptionMessage("Expected to get the 2-nth stream element, but the stream has 0 element(s)");
 
         // when