+++ test/Feature/CleanRegex/Match/group/asInt/_unmatched/MatchGroupIntStreamTest.php
@@ -4,7 +4,7 @@ namespace Test\Feature\CleanRegex\Match\group\asInt\_unmatched;
 use PHPUnit\Framework\TestCase;
 use Test\Utils\ExactExceptionMessage;
 use Test\Utils\Functions;
-use TRegx\CleanRegex\Exception\NoSuchElementFluentException;
+use TRegx\CleanRegex\Exception\NoSuchStreamElementException;
 use TRegx\CleanRegex\Exception\SubjectNotMatchedException;
 
 /**
@@ -55,7 +55,7 @@ class MatchGroupIntStreamTest extends TestCase
         $optional = pattern('(Foo)')->match('Bar')->group(1)->asInt()->distinct()->findFirst(Functions::fail());
 
         // then
-        $this->expectException(NoSuchElementFluentException::class);
+        $this->expectException(NoSuchStreamElementException::class);
         $this->expectExceptionMessage('Expected to get group #1 as integer from the first match, but subject was not matched at all');
 
         // when