+++ test/Feature/CleanRegex/Match/offsets/AbstractMatchPatternTest.php
@@ -4,8 +4,8 @@ namespace Test\Feature\TRegx\CleanRegex\Match\offsets;
 use PHPUnit\Framework\TestCase;
 use Test\Utils\CustomSubjectException;
 use Test\Utils\Functions;
-use TRegx\CleanRegex\Exception\NoSuchElementFluentException;
 use TRegx\CleanRegex\Exception\NoSuchNthElementException;
+use TRegx\CleanRegex\Exception\NoSuchStreamElementException;
 use TRegx\CleanRegex\Exception\SubjectNotMatchedException;
 use function pattern;
 
@@ -46,7 +46,7 @@ class AbstractMatchPatternTest extends TestCase
     public function shouldThrow_map_first_OnUnmatchedSubject()
     {
         // then
-        $this->expectException(NoSuchElementFluentException::class);
+        $this->expectException(NoSuchStreamElementException::class);
         $this->expectExceptionMessage('Expected to get the first match offset, but subject was not matched');
 
         // when
@@ -59,7 +59,7 @@ class AbstractMatchPatternTest extends TestCase
     public function shouldThrow_map_nth_OnUnmatchedSubject()
     {
         // then
-        $this->expectException(NoSuchElementFluentException::class);
+        $this->expectException(NoSuchStreamElementException::class);
         $this->expectExceptionMessage('Expected to get the 0-nth stream element, but the subject backing the stream was not matched');
 
         // when
@@ -72,7 +72,7 @@ class AbstractMatchPatternTest extends TestCase
     public function shouldThrow_filter_first()
     {
         // then
-        $this->expectException(NoSuchElementFluentException::class);
+        $this->expectException(NoSuchStreamElementException::class);
         $this->expectExceptionMessage('Expected to get the first stream element, but the stream has 0 element(s)');
 
         // when
@@ -85,7 +85,7 @@ class AbstractMatchPatternTest extends TestCase
     public function shouldThrow_filter_nth()
     {
         // then
-        $this->expectException(NoSuchElementFluentException::class);
+        $this->expectException(NoSuchStreamElementException::class);
         $this->expectExceptionMessage('Expected to get the 2-nth stream element, but the stream has 0 element(s)');
 
         // when